Split the specs from week 8 into specs for 8/9
