+++ src/Homework/Week08/Assignment.hs
@@ -1,34 +1,27 @@
 module Homework.Week08.Assignment (
-  Atom(..),
-  SExpr(..),
   first,
   abParser,
   abParser_,
   intPair,
-  intOrUppercase,
-  zeroOrMore,
-  oneOrMore,
-  spaces,
-  ident,
-  parseSExpr
+  intOrUppercase
 ) where
 
 import Homework.Week08.AParser
 import Control.Applicative
 
--- #1A
+-- #1
 first :: (a -> b) -> (a,c) -> (b,c)
 first = undefined
 
 instance Functor Parser where
   fmap = undefined
 
--- #2A
+-- #2
 instance Applicative Parser where
   pure = undefined
   _ <*> _ = undefined
 
--- #3A
+-- #3
 abParser :: Parser (Char, Char)
 abParser = undefined
 
@@ -38,37 +31,11 @@ abParser_ = undefined
 intPair :: Parser [Integer]
 intPair = undefined
 
--- #4A
+-- #4
 instance Alternative Parser where
   empty = undefined
   _ <|> _ = undefined
 
--- #5A
+-- #5
 intOrUppercase :: Parser ()
 intOrUppercase = undefined
-
--- #1B
-zeroOrMore :: Parser a -> Parser [a]
-zeroOrMore = undefined
-
-oneOrMore :: Parser a -> Parser [a]
-oneOrMore = undefined
-
--- #2B
-spaces :: Parser String
-spaces = undefined
-
-ident :: Parser Ident
-ident = undefined
-
--- #3B
-type Ident = String
-
-data Atom = N Integer | I Ident
-  deriving (Eq, Show)
-
-data SExpr = A Atom | Comb [SExpr]
-  deriving (Eq, Show)
-
-parseSExpr :: Parser SExpr
-parseSExpr = undefined
